requirements.txt
Flask==2.2.5
pymongo==4.4.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0